Fnt->DisplayFont encapsulation + rewrite text rendering algorithm to support vectors of fonts
Previously, the Drawing object contained a pointer to the singly linked, intrusive
Fnt pointer list. Changing this to a vector of DisplayFonts, gives a complexity
improvement and automatically manages the lifetime of the Font objects (without
the risk of overflowing stack upon distruction as before). Additionally, copy
construction of Fnt is now disabled to prevent use-after-frees.