Replace enum-indexed color array with ColorScheme struct
C++ doesn't have designated initializers for arrays, this makes the configuration
fragile and less expressive than C. This is fixed by replacing the color array
with a Theme struct (which does support designated initializers). This also allow
the lifetime of color theme to be managed automatically and incorrect configurations
to be identified as part of the type system.