Client object encapsulation: dwm is free of malloc/ free
This commit moves a large section of the dwm core into a 'Client' class.
This ensures the class is always in a valid state, and greatly improves
the readability of the common client operations. This commit does not
attempt to change any of the layout or positioning algorithms. As a result,
the encapsulation is incomplete: the Client is still an internally linked
list and manages information about its containing 'Monitor'.

In future changes, 'Client::next', 'Client::snext', and 'Client::fMonitor'
should be removed and replaced by a dedicated structure. Additionally, when
feasible, 'Client::unmanage' should go in the destructor.