XProperty safety: statically guarantee all XLib parameters are consistent.
This commit adds the file 'x.hpp', ensuring that all XAtoms are correctly
marked as either sentinels or properties. A property can be mutated once
it gets attached to a window whereas a sentinel can only ever be used in
comparison operations (preventing sentinels being mistakenly written eg.
writing to '_NET_WM_STATE_FULLSCREEN' is always a bug).

This also ensures properties cannot be written without first attaching
the property to a window, this prevents errors such as mistakenly writing
to the global window or overwriting properties that should only exist in
the context of a client.